Still working on GeoChatFragment;

+++ client/app/src/main/java/com/qubacy/geoqq/data/geo/message/repository/GeoMessageDataRepository.kt
@@ -26,8 +26,8 @@ class GeoMessageDataRepository @Inject constructor(
 ) : ProducingDataRepository(coroutineDispatcher, coroutineScope), MessageDataRepository {
     suspend fun getMessages(
         radius: Int,
-        longitude: Double,
-        latitude: Double
+        longitude: Float,
+        latitude: Float
     ): GetGeoMessagesDataResult {
         val accessToken = mTokenDataRepository.getTokens().accessToken
